refactor: split two more test god-files into tests/ dirs
Continues the god-file cleanup. Splits two flat 'mod tests;' test files that
exceeded the 1500-line structural threshold into tests/ directories, each with
a mod.rs holding shared setup (imports, helpers, structs/impls, and consts made
pub(super) for the child partitions) plus themed partition files:

- agent_task_promotion/tests.rs (2778 lines) -> tests/ (4 parts, 51 tests, 16 shared)
- upgrade_runners/tests.rs (1700 lines) -> tests/ (3 parts, 26 tests, 18 shared)

All 77 tests preserved (verified counts) and run under the new module paths.
Every partition is now under 1500 lines. mod tests; resolves to tests/mod.rs
automatically.

Pure test reorganization, zero production change.
cargo check --workspace --tests: GREEN.
